move ClarityType enum to constants file
