+++ src/constants.ts
@@ -81,6 +81,26 @@ enum AssetType {
   NonFungible = "02"
 }
 
+/**
+ * Type IDs corresponding to each of the Clarity value types as described here:
+ * {@link https://github.com/blockstack/blockstack-core/blob/sip/sip-005/sip/sip-005-blocks-and-transactions.md#clarity-value-representation}
+ */
+enum ClarityType {
+  Int = '00',
+  UInt = '01',
+  Buffer = '02',
+  BoolTrue = '03',
+  BoolFalse = '04',
+  PrincipalStandard = '05',
+  PrincipalContract = '06',
+  ResponseOk = '07',
+  ResponseErr = '08',
+  OptionalNone = '09',
+  OptionalSome = '0a',
+  List = '0b',
+  Tuple = '0c',
+}
+
 export {
   MAX_STRING_LENGTH_BYTES,
   CLARITY_INT_SIZE,
@@ -101,5 +121,6 @@ export {
   FungibleConditionCode,
   NonFungibleConditionCode,
   PrincipalType,
+  ClarityType,
   AssetType
 }
